backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
pdf2image
sqlalchemy
